Move combinators from /util/combinators/* to /util/*
