--- tower/src/util/combinators/mod.rs
@@ -1,9 +1,0 @@
-mod map_err;
-mod map_ok;
-mod try_with;
-mod with;
-
-pub use map_err::{MapErr, MapErrLayer};
-pub use map_ok::{MapOk, MapOkLayer};
-pub use try_with::{TryWith, TryWithLayer};
-pub use with::{With, WithLayer};
